+++ src/components/Navbar.jsx
@@ -40,7 +40,7 @@ const Navbar = () => {
             <nav className="navbar">
                 <div className="container navbar-container">
                     <Link to="/" className="navbar-logo" onClick={closeMenu}>
-                        CODA
+                        TONARE
                     </Link>
 
                     <button